backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.6.0
pydantic[email]
pydantic-settings==2.1.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
qrcode==7.4.2
httpx==0.24.1
bcrypt==3.2.2
sendgrid==6.9.7